+++ README.md
@@ -137,6 +137,8 @@ Distributed under the mit License. See [LICENSE.txt](LICENSE.txt) for more infor
 <!-- CONTACT -->
 
 ## Changelog
+v2.1.0
+- Updated initializer to offer project setup for sfc based project or standard project
 v2.0.1
 - Removed erroneous old boilerplate content
 v2.0.0
