+++ lib/genDirectories.js
@@ -1,5 +1,5 @@
 const fs = require('fs/promises');
-const templates = require('./templates');
+const allTemplates = require('./templates');
 
 const templateDirectories = (obj,path) => {
   return Object.entries(obj)
@@ -70,6 +70,7 @@ const genDirectories = async (answers) => {
     sheetJSON.compendium = answers.compendiumName;
   }
   const relDirectory = `./${answers.directoryName}`.replace(/\/$/g,'');
+  const templates = allTemplates[answers.architecture];
   await Promise.all([
     // Live directory and files
     fs.mkdir(`${relDirectory}`,{recursive:true}),
