+++ lib/query.js
@@ -33,6 +33,21 @@ const questions = [
     type: 'input',
     name: 'compendiumName',
     message: 'Associated compendium (Blank for none)'
+  },
+  {
+    type:'list',
+    name:'architecture',
+    message:'Initialize project for single file component use?',
+    choices:[
+      {
+        name:'Yes',
+        value:'sfc'
+      },
+      {
+        name:'No',
+        value:'standard'
+      }
+    ]
   }
 ];
