+++ lib/templates/sfc/components.pug
@@ -1,0 +1,2 @@
+//- Include your component pug files here.
+//- include ./example-include.js
