+++ lib/templates/sfc/javascript.pug
@@ -1,0 +1,2 @@
+//- Include your generic javascript files that are not associated with a specific component/view here.
+//- include ./example-include.js
