+++ lib/templates/sfc/rolltemplate.pug
@@ -1,0 +1,3 @@
+//- If your project has multiple roll templates, include them here along with any generic styling using the scss mixin.
+//- include ./example-include.js
+//- If your project has only one roll template, simply write it here.
