+++ lib/templates/sfc/sections.pug
@@ -1,0 +1,2 @@
+//- Include your section pug files here.
+//- include ./example-include.js
