+++ lib/templates/sfc/sheet-name.pug
@@ -1,5 +1,6 @@
 include k-scaffold
-include ./mixins/_index.pug
+include ./components/_index.pug
+include ./sections/_index.pug
 - const templateName = '{{sheet-name}}';
 //- sheet_state will contain which display of the sheet is currently enabled
 +hidden({name:'sheet version',value:0})
@@ -7,11 +8,11 @@ include ./mixins/_index.pug
 main#main
   +tabs({name:'{{sheet-name}}',defaultActiveTab:'settings'}).flex-box.tiny-gap
     +tab({name:'settings',container:'article'})
-      include articles/_settings.pug
+      include ./views/_index.pug
     //- Add additional tabs as needed
 //- An input to contain template fields that should always be output with a roll.
 +hidden({name:'template start',value:`@{whisper}&{template:${templateName}} {{character_name=@{character_name}}} {{character_id=@{character_id}}}`})
 include rolltemplate/_index.pug
 +kscript('{{sheet-name}}')
-  //- All additional javascript files should start here
+  //- All additional javascript files should be included in the index file
   include ./javascript/_index.pug
