+++ lib/templates/sfc/views.pug
@@ -1,0 +1,2 @@
+//- Include your view pug files here
+//- include ./example-include.js
