+++ lib/templates/standard/components.scss
