+++ lib/templates/standard/fonts.scss
@@ -1,0 +1,1 @@
+// Import your google fonts here
