+++ lib/templates/standard/general.scss
@@ -39,4 +39,7 @@ body{
       }
     }
   }
+  &.sheet-darkmode{
+    @include variables.dark;
+  }
 }
