+++ lib/templates/standard/index.scss
