+++ lib/templates/standard/javascript.pug
@@ -1,0 +1,2 @@
+//- Include your javascript files here.
+//- include ./example-include.js
