+++ lib/templates/standard/mixins.pug
@@ -1,0 +1,2 @@
+//- Include your pug mixin files here.
+//- include ./example-include.pug
