+++ lib/templates/standard/rolltemplate.pug
@@ -1,3 +1,3 @@
-+rolltemplate(templateName)
++templateWrapper(templateName)
   //- Create your template here
   +characterLink
