+++ lib/templates/standard/sheet-name.pug
@@ -1,0 +1,17 @@
+include k-scaffold
+include ./mixins/_index.pug
+- const templateName = '{{sheet-name}}';
+//- sheet_state will contain which display of the sheet is currently enabled
++hidden({name:'sheet version',value:0})
+//- Now, for our actual sheet. We'll wrap this in a main element and give it an ID of main.
+main#main
+  +tabs({name:'{{sheet-name}}',defaultActiveTab:'settings'}).flex-box.tiny-gap
+    +tab({name:'settings',container:'article'})
+      include articles/_settings.pug
+    //- Add additional tabs as needed
+//- An input to contain template fields that should always be output with a roll.
++hidden({name:'template start',value:`@{whisper}&{template:${templateName}} {{character_name=@{character_name}}} {{character_id=@{character_id}}}`})
+include rolltemplate/_index.pug
++kscript('{{sheet-name}}')
+  //- All additional javascript files should be included in the index file
+  include ./javascript/_index.pug
