+++ lib/templates/standard/sheet-name.scss
