+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kurohyou/k-init",
-  "version": "2.0.1",
+  "version": "2.1.0",
   "description": "<div id=\"top\"></div>\r <span align=\"center\">",
   "main": "index.js",
   "scripts": {
